[FLINK-12965][table][hive] unify catalog view implementations
